Move compareStreaming and resolvePermissionNumber to Util
