+++ src/struct/ClientUtil.js
@@ -1,5 +1,3 @@
-const { Permissions } = require('discord.js');
-
 /**
  * Client utilities to help with common tasks.
  * @param {AkairoClient} client - The client.
@@ -318,37 +316,6 @@ class ClientUtil {
 
         if (!wholeWord) return name.includes(text);
         return name === text;
-    }
-
-    /**
-     * Resolves a permission number and returns an array of permission names.
-     * @param {number} number - The permissions number.
-     * @returns {string[]}
-     */
-    resolvePermissionNumber(number) {
-        const resolved = [];
-
-        for (const key of Object.keys(Permissions.FLAGS)) {
-            if (number & Permissions.FLAGS[key]) resolved.push(key);
-        }
-
-        return resolved;
-    }
-
-    /**
-     * Compares two member objects presences and checks if they stopped or started a stream or not.
-     * Returns `0`, `1`, or `2` for no change, stopped, or started.
-     * @param {GuildMember} oldMember - The old member.
-     * @param {GuildMember} newMember - The new member.
-     * @returns {number}
-     */
-    compareStreaming(oldMember, newMember) {
-        const s1 = oldMember.presence.activity && oldMember.presence.activity.type === 'STREAMING';
-        const s2 = newMember.presence.activity && newMember.presence.activity.type === 'STREAMING';
-        if (s1 === s2) return 0;
-        if (s1) return 1;
-        if (s2) return 2;
-        return 0;
     }
 }
